fixings for mysql

+++ Reports/database/migrations/2026_06_19_144506_create_teacher_infos_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('teacher_infos', function (Blueprint $table) {
         $table->increments('id');
-        $table->integer('supervisor_id');
+        $table->unsignedBigInteger('supervisor_id');
         $table->string('Teacher_Name');
         $table->integer('Teacher_id')->unique();
         $table->integer('school_id');
